fixed site ID duplication error
The code for combining fish and macroinvertebrates (3_combine_fish_macros.R) had an error ~ line78. The problem was that we joined by year_month, but only one of the datasets had that column. As a result, the right_join duplicated everything in the first data set. That is now fixed by adding the year_month column to all data sets before joining.